supergroup support
* update user id to int64, since telegram said it "not exceeding 1e13 by
  absolute value"
* add possible chat type in document
* support new fields in telegram.Message

+++ manager_test.go
@@ -9,7 +9,7 @@ import (
 
 func makeTestUser(name string) *telegram.User {
 	return &telegram.User{
-		ID:        rand.Int(),
+		ID:        int64(rand.Int()),
 		FirstName: name,
 	}
 }
